pyproject.toml
[project]
name = "myteam"
version = "0.2.16"
description = "Agent roster CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fire>=0.5",
    "requests>=2.32.5",
    "pyyaml>=6.0",
]
authors = [{ name = "BeanLab" }]
license = { file = "LICENSE" }

[project.scripts]
myteam = "myteam.cli:main"

[dependency-groups]
dev = [
    "faster-whisper>=1.2.1",
    "pytest>=8.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["uv-build>=0.2"]
build-backend = "uv_build"
